chat_service/pyproject.toml
[project]
name = "chat-service"
version = "0.1.0"
description = "ASTU Connect Chat Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-jose[cryptography]",
    "pydantic-settings",
    "aiokafka",
    "redis",
    "cassandra-driver",
    "websockets",
]
